+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [1.0.6] - 
 
-- fixed ReactiveCheckbox exception when binding with FormControl with null value
-- more tests to Reactive Forms added
+- fixed ReactiveCheckbox and ReactiveCheckboxListTile exception when binding 
+  with FormControl with null value.
+- more tests to Reactive Forms added.
 
 ## [1.0.5] - 2020-07-10